Change fQlen and contructor input of AutoCorrelationCalc to be size_t

+++ src/optimise/AutoCorrelationCalc.h
@@ -5,10 +5,10 @@
 
 class AutoCorrelationCalc{
  public:
- AutoCorrelationCalc(int nVals_) : fDiffs(std::vector<double>(nVals_, 0)),
-                                   fNorms(std::vector<double>(nVals_, 0)),
-                                   fQLen(nVals_)
-                                   {}
+ AutoCorrelationCalc(size_t nVals_) : fDiffs(std::vector<double>(nVals_, 0)),
+                                      fNorms(std::vector<double>(nVals_, 0)),
+                                      fQLen(nVals_)
+                                      {}
 
     void Fill(double);
     double Mean() const;
@@ -21,6 +21,6 @@ class AutoCorrelationCalc{
     std::deque<double>    fValues;
     std::vector<double>   fDiffs;
     std::vector<double>   fNorms;
-    int fQLen;
+    size_t fQLen;
 };
 #endif
